Enable compatibility with Rust 1.12
Rayon currently supports Rust 1.12 or later, but we can't update to the
latest coco because of a small difference in item visibility to `use`.
It's a simple change to make this work though.

There's also a difference that the old `Ordering` didn't have the hidden
`__Nonexhaustive` variant, but we can deal with this by just matching
`SeqCst` in the catch-all pattern too.